planner_backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
distro==1.9.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.7.1
openai==1.55.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1

